solving task 11

+++ 0x1E-search_algorithms/104-advanced_binary.c
@@ -14,7 +14,7 @@ int advanced_binary(int *array, size_t size, int value)
 {
 	size_t left;
 	size_t right;
-	size_t result;
+	int result;
 	size_t i;
 	int mid;
